# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/projects/{project_id}/branches": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "get all branches",
                "tags": [
                    "branch"
                ],
                "summary": "get all project branches",
                "operationId": "get-branches",
                "parameters": [
                    {
                        "type": "string",
                        "description": "project id",
                        "name": "project_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseBranches"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/presenter.ResponseError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "create a branch",
                "tags": [
                    "branch"
                ],
                "summary": "create a branch",
                "operationId": "create-branch",
                "parameters": [
                    {
                        "type": "string",
                        "description": "project id",
                        "name": "project_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "creation request body",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.RequestBranchCreate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseBranch"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/presenter.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/v1/projects/{project_id}/branches/{branch_name}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "get a project branch",
                "tags": [
                    "branch"
                ],
                "summary": "get a project branch",
                "operationId": "get-branch",
                "parameters": [
                    {
                        "type": "string",
                        "description": "project id",
                        "name": "project_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "branch name",
                        "name": "branch_name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseBranch"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/presenter.ResponseError"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "update a branch",
                "tags": [
                    "branch"
                ],
                "summary": "update a branch",
                "operationId": "update-branch",
                "parameters": [
                    {
                        "type": "string",
                        "description": "project id",
                        "name": "project_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "branch name",
                        "name": "branch_name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.ResponseBranch"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/presenter.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/v1/projects/{project_id}/branches/{branch_name}/pipelines/{pipeline_name}": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "create a pipeline",
                "tags": [
                    "pipeline"
                ],
                "summary": "create a pipeline",
                "operationId": "create-pipeline",
                "parameters": [
                    {
                        "type": "string",
                        "description": "project id",
                        "name": "project_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "branch name",
                        "name": "branch_name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "pipeline name",
                        "name": "pipeline_name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/presenter.ResponseError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controller.RequestBranchCreate": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "maxLength": 64,
                    "minLength": 3
                }
            }
        },
        "controller.ResponseBranch": {
            "type": "object"
        },
        "controller.ResponseBranches": {
            "type": "object"
        },
...
        "presenter.ResponseError": {
            "type": "object",
            "properties": {
                "error": {}
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "\"Type 'Bearer TOKEN' to correctly set the API Key\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}